routes: Implement support for categories
This changes the code so that in the routes we now build up the
repository path from the category and the name. If no category is
provided you simply end up with the name. This path is now used
everywhere a path to a repository is constructed which takes care of 80%
of making categories work. Handlers on the mux are added for the
category matching and reordered to ensure we hit the right entry first.

The new repository struct contains all info we need for a repository and
we default to allowing some fields to be empty if we can't retrieve the
information whereas previously we might've returned an error.

The entries is a single level tree of sorts that lifts the category, if
available, to be its name and otherwise uses the repo name. We can use
this to sort all entries alphabetically and now loop over them in our
template.